+++ core-ng/src/main/java/core/framework/impl/web/response/BeanBody.java
@@ -1,11 +1,13 @@
 package core.framework.impl.web.response;
 
-import core.framework.impl.log.filter.BytesParam;
+import core.framework.impl.log.filter.JSONParam;
 import io.undertow.io.Sender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.ByteBuffer;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
  * @author neo
@@ -21,7 +23,7 @@ public final class BeanBody implements Body {
     @Override
     public void send(Sender sender, ResponseHandlerContext context) {
         byte[] body = context.responseBeanMapper.toJSON(bean);
-        logger.debug("[response] body={}", new BytesParam(body));
+        logger.debug("[response] body={}", new JSONParam(body, UTF_8));
         sender.send(ByteBuffer.wrap(body));
     }
 }