+++ core-ng/src/test/java/core/framework/impl/http/HTTPClientImplTest.java
@@ -11,11 +11,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * @author neo
@@ -32,28 +32,29 @@ class HTTPClientImplTest {
     void responseBodyWithNoContent() throws IOException {
         byte[] body = httpClient.responseBody(null);   // apache http client return null for HEAD/204/205/304
 
-        assertEquals("", new String(body, StandardCharsets.UTF_8));
+        assertThat(new String(body, UTF_8)).isEqualTo("");
     }
 
     @Test
     void parseHTTPStatus() {
-        assertEquals(HTTPStatus.OK, HTTPClientImpl.parseHTTPStatus(200));
+        assertThat(HTTPClientImpl.parseHTTPStatus(200)).isEqualTo(HTTPStatus.OK);
     }
 
     @Test
     void parseUnsupportedHTTPStatus() {
-        assertThrows(HTTPClientException.class, () -> HTTPClientImpl.parseHTTPStatus(525));
+        assertThatThrownBy(() -> HTTPClientImpl.parseHTTPStatus(525))
+                .isInstanceOf(HTTPClientException.class);
     }
 
     @Test
     void httpRequest() {
-        HTTPRequest request = new HTTPRequest(HTTPMethod.POST, "http://localhost/uri");
+        var request = new HTTPRequest(HTTPMethod.POST, "http://localhost/uri");
         request.addParam("query", "value");
         request.accept(ContentType.APPLICATION_JSON);
         request.body("text", ContentType.TEXT_PLAIN);
 
         HttpUriRequest httpRequest = httpClient.httpRequest(request);
-        assertEquals("http://localhost/uri?query=value", httpRequest.getURI().toString());
-        assertEquals(ContentType.APPLICATION_JSON.toString(), httpRequest.getFirstHeader(HTTPHeaders.ACCEPT).getValue());
+        assertThat(httpRequest.getURI().toString()).isEqualTo("http://localhost/uri?query=value");
+        assertThat(httpRequest.getFirstHeader(HTTPHeaders.ACCEPT).getValue()).isEqualTo(ContentType.APPLICATION_JSON.toString());
     }
 }