+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Change log
-### 6.8.0 (9/5/2018 - )  !!! api cleanup, remove replaceable method  !!!
+### 6.8.0 (9/5/2018 - 9/6/2018)  !!! api cleanup, remove replaceable method  !!!
 * session: removed session.remove(key), just use session.set(key, null)
 * redis: log returned value
 * validator: log bean if failed to validate, to help troubleshooting
 * http: removed module config, route(), replaced with http().route(method,path,controller)
+* http: mask json response, beanBody/httpClient
 
 ### 6.7.1 (8/31/2018 - 9/5/2018)
 * http: set tcp keep alive timeout to 620s, to adapt to both AWS ELB and gcloud LB